cloud/functions/fetch_prices/requirements.txt
functions-framework==3.*
yfinance>=0.2.31
google-cloud-storage>=2.14.0
